Began Fire Explosion Spell Functionality
- Created Player, Enemy and DestructibleObjects layers.
- Added SpellManager child object with associated script to Player in Test_PlayerController scene.
- Updated PlayerController script to use the current spell through the SpellManager script when the UseSpell input is pressed.
- Added Spell Manager, cast cooldown and damage variables to Spell base script.
- Created FireExplosionSpell script that inherits from the Spell script and summons a fire explosion.
- Created FireExplosion script that damages all enemies in its collision area.
- Temporarily updated SpellManager script to store the mana data in itself rather than the PlayerCharacter SO.
- Created red transparent FireExplosion material for testing purposes.
- Created FireExplosion prefab with a sphere mesh, sphere collider and FireExplosion script.
- Updated SpellManager script to use the current spell's cast cooldown and only cast the spell if the cooldown has ended.
- Created Damageable script that takes a damage amount and evokes an onDamaged event.
- Created Test Dummy script that stores the dummy's health and takes damage when the onDamaged event occurs, dying if its health reaches 0.
- Added a test dummy with a Rigidbody, BoxCollider, TestDummy and Damageable component to the Test_PlayerController scene.

+++ Assets/Scripts/PlayerCharacter.cs
@@ -80,12 +80,12 @@ public class PlayerCharacter : ScriptableObject
     //-------------------------------------------------------------
     // MANA
     //-------------------------------------------------------------
-    public void UseMana(float amount)
+    public void UseMana(int amount)
     {
         curMana -= amount;
     }
 
-    public void GainMana(float amount)
+    public void GainMana(int amount)
     {
         curMana += amount;
     }
